requirements.txt
PyTelegramBotAPI
requests
sentence-transformers
scipy
numpy
psycopg2-binary
pandas
openpyxl